requirements.txt
torch==2.3.1
torchvision==0.18.1
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy==1.26.4
paho-mqtt==1.6.1
